importer/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "parana-importer"
version = "0.1.0"
description = "JaCoCo XML importer for the Parana coverage-tracking system"
requires-python = ">=3.12"
dependencies = [
    "psycopg[binary]>=3.1",
    "lxml>=5.0",
    "click>=8.1",
    "python-dotenv>=1.0",
    "gitpython>=3.1",
]

[project.scripts]
parana-import = "parana_importer.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/parana_importer"]

[tool.hatch.build.targets.wheel.sources]
"src" = ""

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "testcontainers[postgres]>=4.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
